Moved store pieces to individual files before it got messy
