+++ src/js/store/index.js
@@ -1,0 +1,15 @@
+import Vue from 'vue/dist/vue';
+import Vuex from 'vuex';
+import actions from './actions';
+import mutations from './mutations';
+import state from './state';
+
+Vue.use(Vuex);
+
+const store = new Vuex.Store({
+  actions,
+  mutations,
+  state,
+});
+
+export default store;
