+++ src/js/store/state.js
@@ -1,0 +1,4 @@
+export default {
+  isRecording: false,
+  tracks: [],
+};
